update member mapper and add member/team repos

+++ src/internal/repository/abstract/member.go
@@ -8,5 +8,5 @@ import (
 type IMemberRepository interface {
 	Create(conn abstract.IDBConnection, member *domain.Member) error
 	GetByTelegramID(conn abstract.IDBConnection, telegramUserID int64) (*domain.Member, error)
-	GetByTeamID(conn abstract.IDBConnection, teamID int64) ([]*domain.Member, error)
+	GetByTeamID(conn abstract.IDBConnection, teamID int) ([]*domain.Member, error)
 }
