+++ src/internal/repository/abstract/team.go
@@ -7,5 +7,5 @@ import (
 
 type ITeamRepository interface {
 	Create(conn abstract.IDBConnection, team *domain.Team) error
-	GetByID(conn abstract.IDBConnection, id int64) (*domain.Team, error)
+	GetByID(conn abstract.IDBConnection, id int) (*domain.Team, error)
 }
